zuul-core: enable plaintext body caching (#1008)
Despite the name, this is actuall for metatron body caching, when it is optimistically enabled.   The caching needs to happen in ProxyEndpoint, because a server has not yet been decided, and we dont know if the cached copy is needed.    In the case of retries, the cached copy only gets stored when there are multiple failures, indicating that the *second* attempt should be retried.

This has been stable in one region for a few weeks, and is ready for being default on.